+++ bonusPrintf/libft/libft.h
@@ -69,7 +69,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(long long n, int fd);
+void	ft_putnbr_fd(int n, int fd);
 
 /*		linked list		*/
 t_list	*ft_lstnew(void *content);
